playground: extract setupTimeRangeSlider to its own module
The slider's pointer/keyboard wiring + thumb math was 160 lines stuck
in the middle of main.js. Adding the actual-layout-box reads pushed
main.js past the 600-line hard cap, so move the whole thing to
main/time-range-slider.js — a sibling module that imports the same
state / fetcher / draw helpers main.js was already pulling in. main.js
drops to 437 lines and stops importing the live-history fetcher,
graphRange, and setGraphRange (now consumed only by the slider module).

No behavior change beyond the previous commit.
